+++ code/libraries/SIPMoistureSensors/moisture_sensors.h
@@ -10,7 +10,7 @@ class TestSensor : public MoistureSensor{
   // A test class assuming a simple analog sensor.
   private:
     int m_pin; // Arduino pin number.
-    DataStream<int>* m_data;
+    DataStream<float>* m_data;
   
   public:
     TestSensor(int input_pin);
@@ -22,7 +22,7 @@ class WatermarkSensor : public MoistureSensor{
   // Controls the Watermark Sensor (Gypsum block).
 
   private:
-    DataStream<int>* m_data;
+    DataStream<float>* m_data;
   
   public:
     WatermarkSensor(int input_pin);
